requirements.txt
click==8.0.1
colorama==0.4.4
dataclasses
Flask==2.0.1
gunicorn==20.1.0
imgcompare==2.0.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.2
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
